requirements.txt
# -------------------------------
# Core AI / ML & Embeddings
# -------------------------------
torch>=2.2.2,<3.0
transformers>=4.42.2,<5.0
sentence-transformers>=3.1.1,<4.0
ctransformers>=0.2.27,<1.0  # For local LLM inference 

# -------------------------------
# Vector Database
# -------------------------------
#pinecone>=6.0.0,<7.0
pinecone
#pinecone-client>=6.0.0,<7.0
langchain-pinecone==0.1.3



# -------------------------------
# LLMs & LangChain
# -------------------------------
groq
langchain>=0.2.14,<0.3
langchain-community>=0.0.200
langchain_huggingface

# -------------------------------
# PDF & File Utilities
# -------------------------------
pypdf>=3.10.0,<4.0
tqdm
numpy>=1.26,<3.0
scikit-learn>=1.5.2,<2.0
python-dotenv

# -------------------------------
# Web Framework
# -------------------------------
flask>=3.0.3,<4.0

# -------------------------------
# Local Package
# -------------------------------
-e .
